+++ src/index.ts
@@ -7,12 +7,81 @@ import * as dataRoot from "./data-root.js"
 import * as pathing from "./path.js"
 import * as ucan from "./ucan/internal.js"
 
-import { InitOptions, InitialisationError } from "./init/types.js"
-import { State, scenarioContinuation, scenarioNotAuthorised, validateSecrets } from "./auth/state.js"
+import { AppInitOptions, InitialisationError, PermissionedAppInitOptions } from "./init/types.js"
+import { Permissions } from "./ucan/permissions.js"
+import { validateSecrets } from "./auth/state.js"
 import { bootstrapFileSystem, loadFileSystem } from "./filesystem.js"
 
 import FileSystem from "./fs/index.js"
 
+import { isPermissionedAppState } from "./auth/state.js"
+import { setImplementations } from "./setup.js"
+import { BASE_IMPLEMENTATION } from "./auth/implementation/base.js"
+import { USE_WNFS_IMPLEMENTATION } from "./auth/implementation/use-wnfs.js"
+import { LOBBY_IMPLEMENTATION } from "./auth/implementation/lobby.js"
+import { AppState } from "./auth/state/app.js"
+import { PermissionedAppState } from "./auth/state/permissionedApp.js"
+import * as appState from "./auth/state/app.js"
+import * as permissionedAppState from "./auth/state/permissionedApp.js"
+
+
+/**
+ * Check if we're authenticated and initiate the user's file system if 
+ * authenticated (can be disabled).
+ *
+ * See `loadFileSystem` if you want to load the user's file system yourself.
+ * NOTE: Only works on the main/ui thread, as it uses `window.location`.
+ *
+ */
+export async function app(options: AppInitOptions): Promise<AppState> {
+  options = options || {}
+
+  const { useWnfs = false } = options
+
+  if (useWnfs) {
+    setImplementations(USE_WNFS_IMPLEMENTATION)
+  } else {
+    setImplementations(BASE_IMPLEMENTATION)
+  }
+
+  // Check if browser is supported
+  if (globalThis.isSecureContext === false) throw InitialisationError.InsecureContext
+  if (await isSupported() === false) throw InitialisationError.UnsupportedBrowser
+
+  const authedUsername = await common.authenticatedUsername()
+
+  if (authedUsername && useWnfs) {
+    const dataCid = navigator.onLine ? await dataRoot.lookup(authedUsername) : null // data root on server or DNS
+    const logCid = await cidLog.newest() // data root in browser
+    const rootPermissions = { fs: { private: [pathing.root()], public: [pathing.root()] } }
+
+    if (dataCid === null && logCid === undefined) {
+      return appState.scenarioAuthed(
+        authedUsername,
+        await bootstrapFileSystem(rootPermissions)
+      )
+    } else {
+      const fs = options.loadFileSystem === false ?
+        undefined :
+        await loadFileSystem(rootPermissions, authedUsername)
+
+      return appState.scenarioAuthed(
+        authedUsername,
+        fs
+      )
+    }
+
+  } else if (authedUsername) {
+    return appState.scenarioAuthed(
+      authedUsername,
+      undefined
+    )
+
+  } else {
+    return appState.scenarioNotAuthed()
+
+  }
+}
 
 /**
  * Check if we're authenticated, process any lobby query-parameters present in the URL,
@@ -21,15 +90,15 @@ import FileSystem from "./fs/index.js"
  * See `loadFileSystem` if you want to load the user's file system yourself.
  * NOTE: Only works on the main/ui thread, as it uses `window.location`.
  *
- * The `useWnfs` option is only necessary for apps that create their own WNFS. Apps that
- * request filesystem `permissions` from the auth lobby can ignore `useWnfs` in this version
- * of webnative.
  */
-export async function initialise(options: InitOptions): Promise<State> {
+export async function permissionedApp(options: PermissionedAppInitOptions): Promise<PermissionedAppState> {
   options = options || {}
 
   const permissions = options.permissions || null
-  const { useWnfs = false, rootKey } = options
+  const { rootKey } = options
+
+
+  setImplementations(LOBBY_IMPLEMENTATION)
 
   const maybeLoadFs = async (username: string): Promise<undefined | FileSystem> => {
     return options.loadFileSystem === false
@@ -45,7 +114,7 @@ export async function initialise(options: InitOptions): Promise<State> {
   const state = await auth.init(options)
 
   // Allow auth implementation to return a State directly
-  if (state) {
+  if (state && isPermissionedAppState(state)) {
     return state
   }
 
@@ -56,56 +125,35 @@ export async function initialise(options: InitOptions): Promise<State> {
     const validUcans = ucan.validatePermissions(permissions, authedUsername)
 
     if (validSecrets && validUcans) {
-      return scenarioContinuation(
+      return permissionedAppState.scenarioContinuation(
         permissions,
         authedUsername,
         await maybeLoadFs(authedUsername)
       )
     } else {
-      return scenarioNotAuthorised(permissions)
-    }
-
-  } else if (authedUsername && useWnfs) {
-    const dataCid = navigator.onLine ? await dataRoot.lookup(authedUsername) : null // data root on server or DNS
-    const logCid = await cidLog.newest() // data root in browser
-    const rootPermissions = { fs: { private: [pathing.root()], public: [pathing.root()] } }
-
-    if (dataCid === null && logCid === undefined) {
-      return scenarioContinuation(
-        rootPermissions,
-        authedUsername,
-        await bootstrapFileSystem(rootPermissions)
-      )
-    } else {
-      const fs = options.loadFileSystem === false ?
-        undefined :
-        await loadFileSystem(rootPermissions, authedUsername)
-
-      return scenarioContinuation(
-        rootPermissions,
-        authedUsername,
-        fs
-      )
+      return permissionedAppState.scenarioNotAuthorised(permissions)
     }
 
   } else if (authedUsername) {
-    return scenarioContinuation(
+    return permissionedAppState.scenarioContinuation(
       permissions,
       authedUsername,
       await maybeLoadFs(authedUsername),
     )
 
   } else {
-    return scenarioNotAuthorised(permissions)
+    return permissionedAppState.scenarioNotAuthorised(permissions)
 
+    
   }
 }
 
 
 /**
- * Alias for `initialise`.
+ * Alias for `permissionedApp`.
  */
-export { initialise as initialize }
+ export { permissionedApp as initialise }
+ export { permissionedApp as initialize }
 
 
 
@@ -133,9 +181,11 @@ export * from "./filesystem.js"
 export * from "./common/version.js"
 
 export const fs = FileSystem
-export { Scenario, State } from "./auth/state.js"
-export { AuthCancelled, AuthSucceeded, Continuation, NotAuthorised } from "./auth/state.js"
-export { InitialisationError, InitOptions } from "./init/types.js"
+export { AppScenario, AppState } from "./auth/state/app.js"
+export { AuthCancelled, AuthSucceeded, Continuation, NotAuthorised } from "./auth/state/permissionedApp.js"
+export { Authed, NotAuthed } from "./auth/state/app.js"
+export { PermissionedAppScenario as Scenario, PermissionedAppState as State } from "./auth/state/permissionedApp.js"
+export { AppInitOptions, InitialisationError, PermissionedAppInitOptions } from "./init/types.js"
 
 export * as account from "./auth/index.js"
 export * as apps from "./apps/index.js"