+++ src/init/types.ts
@@ -12,12 +12,23 @@ export enum InitialisationError {
 
 
 // INTIALISE
-export type InitOptions = {
+
+
+export type InitOptions = AppInitOptions | PermissionedAppInitOptions
+
+export type AppInitOptions = {
+  // Options
+  loadFileSystem?: boolean
+  rootKey?: string
+  useWnfs?: boolean
+}
+
+export type PermissionedAppInitOptions = {
   permissions?: Permissions
 
   // Options
   autoRemoveUrlParams?: boolean
   loadFileSystem?: boolean
   rootKey?: string
-  useWnfs?: boolean
 }
+